Fix Square v44: wrap basePriceMoney.amount in BigInt()
v44 requires money amounts as BigInt, not numbers.

+++ server/routes/golf-payments.js
@@ -64,7 +64,7 @@ async function createPaymentLink({ name, amount, metadata, redirectUrl, buyerEma
         name,
         quantity: '1',
         basePriceMoney: {
-          amount:   Math.round(amount * 100),
+          amount:   BigInt(Math.round(amount * 100)),
           currency: 'USD',
         },
       }],
